* Moved entity classes for the rules engine into a separate folder. * Added class for the domain used by the rules engine.
