+++ src/Nethereum.eShop/ApplicationCore/Entities/RulesEngine/RuleTree.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Nethereum.eShop.ApplicationCore.Entities
+namespace Nethereum.eShop.ApplicationCore.Entities.RulesEngine
 {
     public class RuleTree: BaseEntity
     {
