+++ src/Nethereum.eShop/ApplicationCore/Interfaces/IRulesEngineService.cs
@@ -1,11 +1,11 @@
-﻿using Nethereum.eShop.ApplicationCore.Entities;
+﻿using Nethereum.eShop.ApplicationCore.Entities.RulesEngine;
 using System.Threading.Tasks;
 
 namespace Nethereum.eShop.ApplicationCore.Interfaces
 {
     public interface IRulesEngineService
     {
-        Task<RuleTree> CreateRuleTreeAsync(RuleTreeOrigin origin);
+        Task<RuleTree> CreateRuleTreeAsync(RulesDomain domain, RuleTreeOrigin origin);
 
         Task Execute(RuleTree TargetRuleTree);
     }
